refactor : Section, Line Service 통합 및 중복 제거
